Add mirror, speed for YouTube videos
As noted in options.js, you can change video speed in YouTube with
the extension and mirror videos (in Chrome). Still need to check in
Firefox. Also should fix mirroring in iframes (such as in DDG).

Upcoming goal: add keyboard shortcuts for speeding up/down video and
looping sections of the video.