WIP video loop, added args for scripting
Still need to clean up the code and potentially change the mirroring
function to take in a param to reduce redundant code. Committing now
in case I work on another machine tomorrow.

NTS: is the args parameter also in the MDN docs? Or do they guarantee
this? It worked for me, but I'd like to double check since I found it
from the Chrome docs.