Speed CSS done
Reworking the frontend should make the backend a lot easier to implement
and makes me more excited to work on the app, so I think this is
worthwhile. I plan on finishing all of the main page's CSS tomorrow. On
Friday I can work on the background script.

I think I can finish this within 2.5 weeks and publish by the end of the
month. I'll ask my testers
1. How long it takes them to currently learn a dance,
2. How often they learn a dance each month,
3. How many times they feel the urge to learn a dance (and how often
   they follow-through on that urge).
And then ask them a few weeks after using the extension. That'll give me
a good understanding of how useful the extension is.