feat: Parse video current time, icons clickable
The delay and loop icons are now clickable. Also, I was having issues
with the min:sec format with number input (it evaluates to "" even if
the user can write it in). I'll work on making it into a beautiful
custom component later. For now, it'll just be parsed and have no
steps.

I'll work on the loop listener now.