WIP: MutationObserver in the browser in scratchpad
This definitely works in the browser to keep the video mirrored, but I
need to be able to toggle this, which I think requires some message
passing between the background and content script.

Also, I need to implement some way of resetting the stored values when a
new video is in use. It could be just tied to the DM logo. I could also
store the URL and clear the session storage every time the URL changes.

But you still have the issue where the video is mirrored, you open a new
tab and run DM there, come back and the video is still mirrored but the
extension doesn't show that. I don't think that's too bad of a
situation, though.